Fix lower bound
I think I found an issue with the way the lower bound is calculated.
Here is an example:

Say the foreground color is rgb(0,0,255) and the background color is
rgba(255,255,0,0.5). With the existing approach a "closest" color is
calculated as rgb(-255,-255,510) and then clipped to rgb(0,0,255).
The overlayed background color than is rgb(128,128,128) which has a
contrast of 2.2 to the foreground color.

To proof that this is not the actual lower bound I can use rgb(0,0,0) as
backdrop color: The overlayed background color in this case is
rgb(128,128,0) which has a contrast of 2 to the foreground color.

The issue with the existing algorithm is that it clips each color
channel individually. In other words: Only colors with a similar hue are
considered for `closest`. This ignores the fact that hue has a
significant impact on luminence.

The algorithm I propose mostly ignores the indicidual color channels and
instead uses the foreground luminance and the maximum and minimum
luminances of the background (overlayed over white and black
respectively).

Since the luminance function is continuous, we can assume that there
exists a backdrop color that can produce any background luminance
between the minimum and maxiumum values.  There may not be a perfect
match because we are dealing with a discrete color space, but we should
get close enough for all practical considerations.

So if the foreground luminance is somewhere in that range, the lower
bound for contrast is 1 (minimal). If it is not, we use the closest
value for calculating the contrast resulting in `min(onBlack, onWhite)`.